@@ -24,8 +24,8 @@ dependencies:
   cupertino_icons: ^1.0.3
   flutter:
     sdk: flutter
-  stream_chat_flutter: ^9.14.0
-  stream_chat_localizations: ^9.14.0
+  stream_chat_flutter: ^9.15.0
+  stream_chat_localizations: ^9.15.0
 
 flutter:
   uses-material-design: true
