@@ -34,9 +34,9 @@ dependencies:
   lottie: ^3.1.2
   provider: ^6.0.5
   sentry_flutter: ^8.3.0
-  stream_chat_flutter: ^9.14.0
-  stream_chat_localizations: ^9.14.0
-  stream_chat_persistence: ^9.14.0
+  stream_chat_flutter: ^9.15.0
+  stream_chat_localizations: ^9.15.0
+  stream_chat_persistence: ^9.15.0
   streaming_shared_preferences: ^2.0.0
   uuid: ^4.4.0
   video_player: ^2.8.7
